Suppress compiler warnings

+++ source/core/coding/ht_block_decoding_avx2.cpp
@@ -44,10 +44,10 @@ uint8_t j2k_codeblock::calc_mbr(const uint32_t i, const uint32_t j, const uint8_
   uint8_t *state_p1 = block_states + static_cast<size_t>(i + 1) * blkstate_stride + j;
   uint8_t *state_p2 = block_states + static_cast<size_t>(i + 2) * blkstate_stride + j;
 
-  uint8_t mbr0 = state_p0[0] | state_p0[1] | state_p0[2];
-  uint8_t mbr1 = state_p1[0] | state_p1[2];
-  uint8_t mbr2 = state_p2[0] | state_p2[1] | state_p2[2];
-  uint8_t mbr  = mbr0 | mbr1 | (mbr2 & causal_cond);
+  uint32_t mbr0 = state_p0[0] | state_p0[1] | state_p0[2];
+  uint32_t mbr1 = state_p1[0] | state_p1[2];
+  uint32_t mbr2 = state_p2[0] | state_p2[1] | state_p2[2];
+  uint32_t mbr  = mbr0 | mbr1 | (mbr2 & causal_cond);
   mbr |= (mbr0 >> SHIFT_REF) & (mbr0 >> SHIFT_SCAN);
   mbr |= (mbr1 >> SHIFT_REF) & (mbr1 >> SHIFT_SCAN);
   mbr |= (mbr2 >> SHIFT_REF) & (mbr2 >> SHIFT_SCAN) & causal_cond;
@@ -409,7 +409,7 @@ auto process_stripes_block_dec = [](SP_dec &SigProp, j2k_codeblock *block, const
         state_p[0] |= 1 << SHIFT_PI_;
         bit = SigProp.importSigPropBit();
         //        block->modify_state(refinement_value, bit, i, j);
-        state_p[0] |= bit << SHIFT_REF;
+        state_p[0] |= static_cast<uint8_t>(bit << SHIFT_REF);
         *sp |= bit << pLSB;
         *sp |= bit << (pLSB - 1);  // new bin center ( = 0.5)
       }
